+++ include/LuminoEngine/Graphics/SwapChain.hpp
@@ -96,7 +96,7 @@ public:
     static void resizeBackbuffer(SwapChain* swapChain, int width, int height) { swapChain->resizeBackbuffer(width, height); }
     //static void present(SwapChain* swapChain, GraphicsContext* context) { swapChain->present(context); }
     static int imageIndex(SwapChain* swapChain) { return swapChain->imageIndex(); }
-    static int swapBufferCount(SwapChain* swapChain) { return swapChain->m_backbuffers.size(); }
+    static int swapBufferCount(SwapChain* swapChain) { return static_cast<int>(swapChain->m_backbuffers.size()); }
 };
 
 } // namespace detail
